feat: Use R2DBC for reactive database interaction (#1)
- Update Gradle dependencies:
  - Remove Spring Web and replace with Spring Webflux.
  - Remove Springdoc Web and replace with Springdoc Webflux.
  - Add Kotlin Coroutines + coroutines reactor.
  - Add h2r2dbc for R2DBC driver for H2database.
  - Add Spring Data R2DBC.
- Update all controller and service functions to be suspending functions. Now that we're using both Spring Webflux and Kotlin Coroutines with reactor support, we're able to do this. Essentially, this means that Spring Boot will automatically launch a coroutine each time a request is sent.
- Add `CustomerEntity` to represent the customer in the database.
- Update `CustomerRepo` to extend `CoroutineCrudRepository`. This is a Spring Data R2DBC CRUD repository with coroutine support which adds a lot of useful functions by default. Additionally, we can use our IDE to easily auto-complete new custom functions within this class. 
- Add `CustomerRepoFacade` which is a facade that will be used in `CustomerService`. The benefit here is that `CustomerService` does not need to implement the Spring Data R2DBC dependency. It simply needs to send model classes to `CustomerRepoFacade` and then the facade class will handle converting model classes to entity classes and vice versa.

---

One thing to note about is that `id` is nullable in `CustomerEntity`. If we decided to manually generate an UUID in the code and used the .save() function provided by the `CoroutineCrudRepository`, the code would fail when trying to insert a new entry. If you provide an id (annotated with `@Id`) and call `.save()`, R2DBC thinks that you are trying to update an existing entry. This means, if we want to use the `@Id` annotation, we have to let the database automatically generate the UUID. I have updated the initial Liquibase database changelog to ensure that automatic UUID generation works for both H2 and PostgreSQL databases. This could possibly have been solved differently, but I decided to go with this solution.

---

Now this means that there’s suddenly 5 classes which in some way represent a customer. This could be considered overkill but for a large project, this could easily be the norm. The purpose of this repository is simply to showcase how a project could be structured. 
- `CustomerEntity` - represents a customer in the database.
- `Customer` - Model class with pure kotlin code and uses no dependencies.
- `CustomerInput` - Model class with pure kotlin code and uses no dependencies.
- `CustomerDto` - Data transfer object that Jackson can serialize.
- `CustomerInputDto` - Data transfer object that Jackson can serialize.

The split between classes such as `Customer` and `CustomerInput` is because you don’t need the id of a customer to create one. The id is first generated once it has been added to the database. The benefit here is that we avoid having to send null values for an irrelevant id field when trying to create a new customer. This might be a lot of boilerplate code, but you should always consider what fits your specific needs. The very same functionality could easily be implemented with less customer classes. For this project I have simply decided to split it to separate concerns.